Extract provider auth store into providers crate
