+++ crates/tandem-providers/Cargo.toml
@@ -10,7 +10,10 @@ edition = "2021"
 anyhow = "1"
 async-stream = "0.3"
 async-trait = "0.1"
+base64 = "0.22"
+dirs = "6"
 futures = "0.3"
+keyring = "2"
 reqwest = { version = "0.12", features = ["json", "stream"] }
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
@@ -18,3 +21,6 @@ tokio = { version = "1", features = ["full"] }
 tokio-util = "0.7"
 tracing = "0.1"
 tandem-types = { path = "../tandem-types", version = "0.6.1" }
+
+[dev-dependencies]
+tempfile = "3"
