+++ crates/tandem-providers/src/lib.rs
@@ -1,3 +1,7 @@
+pub mod provider_auth_store;
+
+pub use provider_auth_store::*;
+
 include!("lib_parts/part01.rs");
 include!("lib_parts/part02.rs");
 include!("lib_parts/part03.rs");
